BE ~ FEAT ~ implement audio duration validation

+++ backend/migrations/1698657611774_create-question-answer-histories-table.js
@@ -36,7 +36,7 @@ exports.up = (pgm) => {
       notNull: false,
     },
     duration: {
-      type: 'bigint',
+      type: 'float',
       notNull: false,
     },
     retry_attempt: {
